Moved OAuth2 unified flow to libraries/auth, some minor cleanup and documentation
