+++ extensions/auth/portability-auth-facebook/build.gradle
@@ -17,7 +17,6 @@
 dependencies {
     compile project(':portability-spi-api')
     compile project(':portability-spi-cloud')
-    compile project(':extensions:auth:portability-auth-oauth2')
 
     compile "org.slf4j:slf4j-api:${slf4jVersion}"
     compile "org.slf4j:slf4j-log4j12:${slf4jVersion}"
