--- extensions/auth/portability-auth-facebook/src/main/java/org/datatransferproject/auth/facebook/FacebookOAuthConfig.java
@@ -1,51 +1,0 @@
-/*
- * Copyright 2018 The Data Transfer Project Authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.datatransferproject.auth.facebook;
-
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
-import java.util.List;
-import java.util.Map;
-import org.datatransferproject.auth.oauth2.OAuth2Config;
-
-public class FacebookOAuthConfig implements OAuth2Config {
-
-  @Override
-  public String getServiceName() {
-    return "Facebook";
-  }
-
-  @Override
-  public String getAuthUrl() {
-    return "https://www.facebook.com/v3.1/dialog/oauth";
-  }
-
-  @Override
-  public String getTokenUrl() {
-    return "https://graph.facebook.com/v3.1/oauth/access_token";
-  }
-
-  @Override
-  public Map<String, List<String>> getExportScopes() {
-    return ImmutableMap.of("PHOTOS", ImmutableList.of("user_photos"));
-  }
-
-  @Override
-  public Map<String, List<String>> getImportScopes() {
-    return ImmutableMap.of("PHOTOS", ImmutableList.of("user_photos"));
-  }
-}
