+++ extensions/auth/portability-auth-google/build.gradle
@@ -17,7 +17,7 @@
 dependencies {
     compile project(':portability-spi-api')
     compile project(':portability-spi-cloud')
-    compile project(':extensions:auth:portability-auth-oauth2')
+    compile project(':libraries:auth')
 
     compile("com.google.api-client:google-api-client:${googleApiClient}")
     compile("com.google.apis:google-api-services-calendar:${googleCalendarVersion}")
