+++ extensions/auth/portability-auth-instagram/build.gradle
@@ -16,7 +16,7 @@
 dependencies {
     compile project(':portability-spi-api')
     compile project(':portability-spi-cloud')
-    compile project(':extensions:auth:portability-auth-oauth2')
+    compile project(':libraries:auth')
 
     // logging
     compile("org.slf4j:slf4j-api:${slf4jVersion}")
