+++ extensions/auth/portability-auth-microsoft/build.gradle
@@ -17,7 +17,7 @@
 dependencies {
     compile project(':portability-spi-api')
     compile project(':portability-spi-cloud')
-    compile project(':extensions:auth:portability-auth-oauth2')
+    compile project(':libraries:auth')
 
     compile "com.google.http-client:google-http-client:${googleHttpClientVersion}"
     compile "com.google.http-client:google-http-client-jackson2:${googleHttpClientVersion}"
