+++ libraries/auth/build.gradle
@@ -19,12 +19,8 @@ dependencies {
     compile project(':portability-spi-cloud')
 
     compile("com.google.oauth-client:google-oauth-client:${googleApiClient}")
-    compile "com.google.http-client:google-http-client-jackson2:${googleHttpClientVersion}"
+    compile("com.google.http-client:google-http-client-jackson2:${googleHttpClientVersion}")
 
-    compile "org.slf4j:slf4j-api:${slf4jVersion}"
-    compile "org.slf4j:slf4j-log4j12:${slf4jVersion}"
-
-    testCompile("com.google.truth:truth:${truthVersion}")
-    testCompile("junit:junit:${junitVersion}")
-    testCompile("org.mockito:mockito-all:${mockitoVersion}")
+    compile("org.slf4j:slf4j-api:${slf4jVersion}")
+    compile("org.slf4j:slf4j-log4j12:${slf4jVersion}")
 }
