+++ settings.gradle
@@ -2,7 +2,7 @@ rootProject.name = 'org.datatransferproject'
 
 include ':portability-types-client', ':portability-types-transfer', ':portability-types-common'
 include ':portability-spi-service', ':portability-spi-api', ':portability-spi-cloud', ':portability-spi-transfer'
-include ':libraries:security', ':libraries:config', ':libraries:transfer', 'libraries:logging'
+include ':libraries:security', ':libraries:config', ':libraries:auth', ':libraries:transfer', ':libraries:logging'
 include ':portability-test-utilities'
 include ':portability-api-launcher'
 
@@ -17,9 +17,6 @@ include ':distributions:demo-google-deployment:api', ':distributions:demo-google
 
 // Jar Components
 include ':portability-api', ':portability-transfer', ':portability-bootstrap-vm'
-
-// Auth Extensions
-include ':extensions:auth:portability-auth-oauth2'
 
 // Cloud Extensions
 include ':extensions:cloud:portability-cloud-local', ':extensions:cloud:portability-cloud-google', ':extensions:cloud:portability-cloud-microsoft'
